pyproject.toml
[tool.poetry]
name = "unibet-bet-placer"
version = "0.1.0"
description = ""
authors = ["Nick Martinovic <n.martinovic@criteo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
playwright = "^1.52.0"
python-dotenv = "^1.1.0"
fastapi = "^0.115.12"
pydantic = "^2.11.4"
uvicorn = "^0.34.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-api = "app.api:app"
